Move tests to unique servicex_app_tests module
