+++ servicex_app/servicex_app_test/resources/internal/test_add_file_to_dataset.py
@@ -29,7 +29,7 @@ from unittest.mock import ANY
 
 from pytest import fixture
 from servicex_app import LookupResultProcessor
-from tests.resource_test_base import ResourceTestBase
+from servicex_app_test.resource_test_base import ResourceTestBase
 
 from servicex_app.dataset_manager import DatasetManager
