+++ servicex_app/servicex_app_test/resources/test_web_app_init.py
@@ -27,7 +27,7 @@
 #   OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #
 import pytest
-from tests.resource_test_base import ResourceTestBase
+from servicex_app_test.resource_test_base import ResourceTestBase
 
 
 class TestWebAppInit(ResourceTestBase):
