+++ servicex_app/servicex_app_test/resources/transformation/test_deployment.py
@@ -2,7 +2,7 @@ from unittest.mock import MagicMock
 
 from pytest import fixture
 
-from tests.resource_test_base import ResourceTestBase
+from servicex_app_test.resource_test_base import ResourceTestBase
 
 
 class TestDeploymentStatus(ResourceTestBase):
