+++ servicex_app/servicex_app_test/resources/transformation/test_get_one.py
@@ -1,4 +1,4 @@
-from tests.resource_test_base import ResourceTestBase
+from servicex_app_test.resource_test_base import ResourceTestBase
 
 
 class TestTransformationRequest(ResourceTestBase):
