+++ servicex_app/servicex_app_test/resources/transformation/test_status.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from tests.resource_test_base import ResourceTestBase
+from servicex_app_test.resource_test_base import ResourceTestBase
 
 
 class TestTransformStatus(ResourceTestBase):
