+++ servicex_app/servicex_app_test/resources/users/test_accept_user.py
@@ -1,4 +1,4 @@
-from tests.web.web_test_base import WebTestBase
+from servicex_app_test.web.web_test_base import WebTestBase
 
 
 class TestAcceptUser(WebTestBase):
