+++ servicex_app/servicex_app_test/test_app_init.py
