+++ servicex_app/servicex_app_test/test_docker_repo_adapter.py
